use official actions for release creation
The update-release action is in archive mode/no longer being worked on.
Hence, we move to official actions to improve stability of the setup.
This requires now calling three distinct actions: first delete the previous "latest" release, then creating a new release based on the new commit and then uploading the release assets to that release.